fixing the problem with vite first
