+++ os_in_browser/src/context/windowFileSystem.tsx
@@ -4,37 +4,65 @@ import {
   ParentComponent,
   useContext,
   onMount,
+  Accessor,
 } from "solid-js";
 
 import { FSModule } from "browserfs/dist/node/core/FS";
 
+import Index from "../public.json";
 import { installBFS } from "../utils/installBFS";
 import { DIRECTORY_LIST } from "../utils/constants";
+import BrowserFS, { FileSystemConfiguration } from "browserfs";
 
 type FileSystemContextState = {
-  fs: FSModule | null;
+  fs: Accessor<FSModule | null>;
 };
 
 const FileSystemContext = createContext<FileSystemContextState>({
-  fs: installBFS(),
-} as FileSystemContextState);
+  fs: null,
+} as unknown as FileSystemContextState);
+
+const FileSystemConfig: FileSystemConfiguration = {
+  fs: "MountableFileSystem",
+  options: {
+    "/": {
+      fs: "OverlayFS",
+      options: {
+        readable: {
+          fs: "HTTPRequest",
+          options: Index,
+        },
+        writable: {
+          fs: "IndexedDB",
+          options: {
+            storeName: "fsStore",
+          },
+        },
+      },
+    },
+  },
+};
 
 export const FileSystemProvider: ParentComponent = (props) => {
+  const [fs, setFs] = createSignal<FSModule | null>(null);
   onMount(() => {
-    DIRECTORY_LIST.forEach((dir) => {
-      try {
-        console.log(FileSystemContext.defaultValue.fs?.readdirSync("/home"));
-      } catch (e) {
-        // the dir already exists
-        console.log(e);
-
-        return;
+    BrowserFS.install(window);
+    BrowserFS.configure(FileSystemConfig, (e) => {
+      if (e) {
+        throw e;
       }
+      console.log("BFS installed");
     });
+
+    let fs = BrowserFS.BFSRequire("fs");
+    if (!fs) {
+      setFs(null);
+    }
+    setFs(fs);
   });
 
   return (
-    <FileSystemContext.Provider value={FileSystemContext.defaultValue}>
+    <FileSystemContext.Provider value={{ fs }}>
       {props.children}
     </FileSystemContext.Provider>
   );
